refactor(benchmark): split Python project
